Cleaned up duplicate code and moved analyzer to services
