--- src/backend/app/analyzers.py
@@ -1,24 +1,0 @@
-from pathlib import Path
-from typing import Dict, Any
-
-def analyze_image(path: Path) -> Dict[str, Any]:
-    # Placeholder image analysis - return basic metadata
-    return {"type": "image", "path": str(path), "size": path.stat().st_size}
-
-def analyze_content(path: Path) -> Dict[str, Any]:
-    # Placeholder content analysis (text/word)
-    try:
-        text = path.read_text(errors="ignore")
-        length = len(text)
-    except Exception:
-        length = None
-    return {"type": "content", "path": str(path), "length": length}
-
-def analyze_code(path: Path) -> Dict[str, Any]:
-    # Placeholder code analysis - return simple metrics
-    try:
-        text = path.read_text(errors="ignore")
-        lines = text.count("\n") + 1
-    except Exception:
-        lines = None
-    return {"type": "code", "path": str(path), "lines": lines}
