+++ src/backend/app/views/uploadFolderView.py
@@ -76,8 +76,8 @@ class UploadFolderView(APIView):
             return JsonResponse({"error": "Uploaded file is not a zip archive."}, status=400)
 
         # Import analyzers here to avoid circular import problems at module import time.
-        # The analyzers implementation currently lives at app/analyzers.py
-        analyzers = importlib.import_module("app.analyzers")
+        # The analyzers implementation now lives at app/services/analyzers.py
+        analyzers = importlib.import_module("app.services.analyzers")
 
         results = []
         with tempfile.TemporaryDirectory() as tmpdir:
